remoção de return no delete (desnecessário) aplicação de clean code básico não declare variável se só vai usar uma vez
documentoAproveitamentoRepository.buscarDocumentoAproveitamento(inciso, id).get(0)

+++ src/main/java/com/example/sicapweb/web/controller/ap/concurso/EditalAdmissaoController.java
@@ -134,13 +134,10 @@ public class EditalAdmissaoController {
     @CrossOrigin
     @Transactional
     @DeleteMapping(value = {"/processos/{id}"})
-    public ResponseEntity<?> delete(@PathVariable BigInteger id) {
+    public void delete(@PathVariable BigInteger id) {
         ProcessoAdmissao processoAdmissao = processoAdmissaoRepository.findById(id);
-        if (processoAdmissao!= null) {
-            if(processoAdmissao.getStatus()==1){
-                processoAdmissaoRepository.delete(id);
-            }
-        }
-        return ResponseEntity.noContent().build();
+        if (processoAdmissao!= null && processoAdmissao.getStatus()==1) 
+            processoAdmissaoRepository.delete(id);
+             
     }
 }
